fix datatype for int32
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@27839 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-3.2/drivers/mtd/maps/nvram_linux.h
@@ -1,10 +1,10 @@
 #ifndef __NVRAM_LINUX__
 #define	__NVRAM_LINUX__
 
-//typedef unsigned int	uint;
-typedef unsigned long	uint32;
-typedef unsigned short	uint16;
-typedef unsigned char	uint8;
+//typedef unsigned int  uint;
+typedef unsigned int uint32;
+typedef unsigned short uint16;
+typedef unsigned char uint8;
 #define	u32		uint32
 #define	u16		uint16
 #define	u8		uint8
